pyproject.toml
[project]
name = "agent-ouch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai~=0.4.0", # This means >=0.4.0 and <0.5.0
    "python-dotenv==1.1.0",
]